feat: independent list command builder
